Refactor testing into three separate modules
